backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
joblib
scikit-learn
nltk